蔡明辰，分站配送情况分析

+++ Distribution-management-centre/src/main/java/com/example/distributionmanagementcenter/controller/GoodController.java
@@ -143,7 +143,7 @@ public class GoodController {
         return httpResponseEntity;
     }
     @PostMapping(value = "/getGoodByOrderId")
-    public HttpResponseEntity getGoodByOrderId(@RequestBody Integer id) {
+    public HttpResponseEntity getGoodByOrderId(@RequestBody Long id) {
 
         HttpResponseEntity httpResponseEntity = new HttpResponseEntity();
         try {
